***These may be part of a previous commit, not showing up on GitHub for some reason***
Fixed issue with Room not always displaying for selection specifically if the application was started with a valid valt config, but an invalid room id.
Changed the camera class in axis.py to loop every 10 seconds until it connects to the camera. Previously it just failed and did nothing.
Added an automatic orientation setting that adjusts display to Landscape or portrait mode based on window dimensions.
Changed font scaling to primarily use sp instead of fixed font sizes. This may cause some text display issues, but has not appeared in any testing thus far.
Added an optioned to enable and disable the clock.
Fixed the keyboard positioning on Android devices so it no longer covers the input field.
Fixed an issue with the camera control module crashing on newer versions of android. This appeared to be related to audio and was fixed by specifying the SDL_AUDIODRIVE environment variable.
Determined issues with video scaling were the result of the specified rtsp feed resolution. Resolution must be larger than available space as it will scale down, but not up.
Change room status check to cancel if the room id is invalid.
Check will restart when config is changed.
Fixed an issue with the room name displaying incorrectly if the room id was invalid.
Adjusted most sizing options to scale better with higher resolution devices. Only the portrait mode top section and the about page now have fixed sizes.
Fixed an issue with the about page crashing to application if unable to pull the device mac address.

*** 1/24/2024 ***
CameraControl and Axis class will now automatically reconnect to a camera on failure.
Redesigned buttons so they have a rounded appearance with a shadow.
Removed most image buttons and replaced with new button type.
Added volume control to CameraControl.
Consolidated Zoom and PTZ Step Size adjustment.
Adjusted info page scaling to use 15sp font size.
Set a default timeout for VAPIX calls to 5 seconds, can be adjusted.
Added threading to keypad functions.
Added threading to CameraControl functions.
CameraControl now loads video widget in the background and only displays widget once it is loaded.
Keypad and CameraControl classes converted to use Kivy object properties for non-required or default value variables.