Added a wireless test function to the ROAM menu
Added a wireless icon to show when the ROAM Wireless bridge is connected to the wireless network.

Changed size of images in RoundedShadowButtonWithImage. Image now scales based on the size of the button up to the native size of the image.

Resolved a bug that would occur if the setting menu was open but the valt object had not been created yet.