Added section to settings to specify ROAM Camera IP, Username, and Password
Changed so that if ROAM is detected the Camera Control app will automatically connect to the specific ROAM Camera IP using the given username and password instead of querying VALT for the info.

Disabled select camera button if there is only one camera in a room.

Adjusted rtsp url to not pull audio if audio is disabled in the settings.